separated tests which bootstrap the bridge
